!!IMPORTANT!!: Added Email verification feature
I have added the email verification feature, where when the user creates a new account, the details would be saved in the 'unverified' collection, and a verification link will be sent through the email address entered. When the verification link is clicked, the data would be copied into the 'accounts' collection, and the document will be deleted from the 'unverified' collection. Everything works fine right now...